Extract `FeedLoadeStub` into its own file
